feat: Implement drag-and-drop functionality for tasks in calendar view
- Added useCalendarStyles hook to manage global styles for drag highlights.
- Created TaskItem component to represent individual tasks with drag-and-drop support.
- Refactored CalendarView to integrate DnD context and handle task dragging.
- Updated TaskBoard to support task updates from the calendar view.
- Removed inline style management for drag highlights and centralized it in the new hook.
