+++ package.json
@@ -12,6 +12,7 @@
     "@auth/core": "^0.37.4",
     "@auth/mongodb-adapter": "^3.7.4",
     "@dnd-kit/core": "^6.3.1",
+    "@dnd-kit/modifiers": "^9.0.0",
     "@dnd-kit/sortable": "^10.0.0",
     "@dnd-kit/utilities": "^3.2.2",
     "@heroui/avatar": "^2.2.7",
